SEM15LUISCALLE: Restructuracion de vehiculo
