+++ src/main/java/com/transport/management/entities/ViajeEntity.java
@@ -8,6 +8,7 @@ import lombok.Setter;
 import java.time.LocalDateTime;
 
 import com.transport.management.modules.conductor.ConductorEntity;
+import com.transport.management.modules.vehiculo.VehiculoEntity;
 import com.transport.management.utils.abtractBase.BaseEntity;
 
 @Entity
