+++ src/main/java/com/transport/management/modules/vehiculo/VehiculoRepository.java
@@ -1,8 +1,7 @@
-package com.transport.management.repositories;
+package com.transport.management.modules.vehiculo;
 
 import java.util.List;
 
-import com.transport.management.entities.VehiculoEntity;
 import com.transport.management.utils.abtractBase.BaseRepository;
 
 public interface VehiculoRepository extends BaseRepository<VehiculoEntity> {
